fix(function): fix order in register_number_to_string.
