+++ src/query/functions/src/scalars/arithmetic.rs
@@ -16,17 +16,22 @@
 
 use std::sync::Arc;
 
+use common_expression::types::decimal::Decimal;
 use common_expression::types::decimal::DecimalColumn;
 use common_expression::types::nullable::NullableColumn;
 use common_expression::types::nullable::NullableDomain;
+use common_expression::types::number::Number;
+use common_expression::types::number::NumberType;
 use common_expression::types::number::F64;
-use common_expression::types::number::*;
 use common_expression::types::string::StringColumnBuilder;
 use common_expression::types::AnyType;
+use common_expression::types::ArgType;
 use common_expression::types::DataType;
+use common_expression::types::DecimalDataType;
 use common_expression::types::NullableType;
 use common_expression::types::NumberClass;
 use common_expression::types::NumberDataType;
+use common_expression::types::SimpleDomain;
 use common_expression::types::StringType;
 use common_expression::types::ALL_NUMBER_CLASSES;
 use common_expression::types::ALL_NUMERICS_TYPES;
@@ -41,12 +46,17 @@ use common_expression::vectorize_with_builder_2_arg;
 use common_expression::with_number_mapped_type;
 use common_expression::Column;
 use common_expression::ColumnBuilder;
+use common_expression::EvalContext;
 use common_expression::Function;
 use common_expression::FunctionDomain;
 use common_expression::FunctionProperty;
 use common_expression::FunctionRegistry;
 use common_expression::FunctionSignature;
 use common_expression::Scalar;
+use common_expression::TypeDeserializer;
+use common_io::display_decimal_128;
+use common_io::display_decimal_256;
+use ethnum::i256;
 use lexical_core::FormattedSize;
 use num_traits::AsPrimitive;
 
@@ -454,9 +464,9 @@ fn register_string_to_number(registry: &mut FunctionRegistry) {
 }
 
 pub fn register_number_to_string(registry: &mut FunctionRegistry) {
-    for src_type in ALL_NUMERICS_TYPES {
+    for src_type in ALL_NUMBER_CLASSES {
         with_number_mapped_type!(|NUM_TYPE| match src_type {
-            NumberDataType::NUM_TYPE => {
+            NumberClass::NUM_TYPE => {
                 registry
                     .register_passthrough_nullable_1_arg::<NumberType<NUM_TYPE>, StringType, _, _>(
                         "to_string",
@@ -537,6 +547,88 @@ pub fn register_number_to_string(registry: &mut FunctionRegistry) {
                     },
                 );
             }
+            NumberClass::Decimal128 => {
+                register_decimal_to_string(registry)
+            }
+            NumberClass::Decimal256 => {
+                // already registered in Decimal128 branch
+            }
         });
+    }
+}
+
+fn register_decimal_to_string(registry: &mut FunctionRegistry) {
+    // decimal to string
+    registry.register_function_factory("to_string", |_params, args_type| {
+        if args_type.len() != 1 {
+            return None;
+        }
+
+        let arg_type = args_type[0].clone();
+        if !arg_type.is_decimal() {
+            return None;
+        }
+
+        Some(Arc::new(Function {
+            signature: FunctionSignature {
+                name: "to_string".to_string(),
+                args_type: vec![arg_type.clone()],
+                return_type: StringType::data_type(),
+                property: FunctionProperty::default(),
+            },
+            calc_domain: Box::new(|_args_domain| FunctionDomain::Full),
+            eval: Box::new(move |args, tx| decimal_to_string(args, arg_type.clone(), tx)),
+        }))
+    });
+}
+
+fn decimal_to_string(
+    args: &[ValueRef<AnyType>],
+    from_type: DataType,
+    _ctx: &mut EvalContext,
+) -> Value<AnyType> {
+    let arg = &args[0];
+
+    let mut is_scalar = false;
+    let column = match arg {
+        ValueRef::Column(column) => column.clone(),
+        ValueRef::Scalar(s) => {
+            is_scalar = true;
+            let builder = ColumnBuilder::repeat(s, 1, &from_type);
+            builder.build()
+        }
+    };
+
+    let from_type = from_type.as_decimal().unwrap();
+
+    let result = match from_type {
+        DecimalDataType::Decimal128(_) => {
+            let (buffer, from_size) = i128::try_downcast_column(&column).unwrap();
+
+            let mut builder = StringColumnBuilder::with_capacity(buffer.len(), buffer.len() * 10);
+            for x in buffer {
+                builder.put_str(&display_decimal_128(x, from_size.scale));
+                builder.commit_row();
+            }
+            builder.finish_to_column()
+        }
+
+        DecimalDataType::Decimal256(_) => {
+            let (buffer, from_size) = i256::try_downcast_column(&column).unwrap();
+
+            let mut builder = StringColumnBuilder::with_capacity(buffer.len(), buffer.len() * 10);
+            for x in buffer {
+                builder.put_str(&display_decimal_256(x, from_size.scale));
+                builder.commit_row();
+            }
+            builder.finish_to_column()
+        }
+    };
+
+    if is_scalar {
+        let scalar = result.index(0).unwrap();
+        Value::Scalar(scalar.to_owned())
+    } else {
+        Value::Column(result)
     }
 }
